Use a common exception aware process
